Converted amount type into integer Checked for the validity for entered amounts

+++ IncomeTaxDLL/I80CCD2Deduction.cs
@@ -8,7 +8,7 @@ namespace IncomeTax
 {
     interface I80CCD2Deduction
     {
-        double I80CCD2DeductionAmount
+        int I80CCD2DeductionAmount
         {
             get;
             set;
